camera_service/requirements.txt
Flask
opencv-python
requests
flask-cors
Pillow
dotenv
torch
facenet-pytorch